int -> long long

+++ levkin.dima/S2/main.cpp
@@ -15,7 +15,7 @@ int main(int argc, char* argv[])
       }
     }
     std::istream& in = (argc > 1) ? file : std::cin;
-    levkin::Stack< int > results = levkin::parse(in);
+    levkin::Stack< long long > results = levkin::parse(in);
     while (!results.empty()) {
       std::cout << results.drop() << (results.empty() ? "" : " ");
     }
